% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distract.R
\name{copy}
\alias{copy}
\title{Copy the selected entry link in system cliboard.}
\usage{
copy(entry.n = 1)
}
\arguments{
\item{entry.n}{Number of the entry to copy. Get it from the \code{links} or
the \code{distract} commands.}
}
\value{
the link is copied in system clipboard, and print the copied link.
}
\description{
Given an entry number, \code{copy()} copies the link of the entry in system
clipboard, to be used with your favourite web browser.
}
\examples{
\dontrun{
copy(1)
}
}

