% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distract.R
\name{distract}
\alias{distract}
\title{Get available feeds, or feed of a preselected one.}
\usage{
distract(feed.number = NULL)
}
\arguments{
\item{feed.number}{row number of the desired feed}
}
\value{
List of feeds from the chosen source
}
\description{
\code{distract()} given with no input displays the list
of available feeds. Otherwise, passing the feed number, returns
the feed.
}
\examples{
distract()
distract(1)
}

