% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distract.R
\name{feed.extract.with.error.handling}
\alias{feed.extract.with.error.handling}
\title{Feed reader with a custom error handling.}
\usage{
feed.extract.with.error.handling(url)
}
\arguments{
\item{url}{URL for the feed.}
}
\value{
List with parsed feed, or status: ERR
}
\description{
Try to parse the url, and if it encounters an error, returns a list with an
unique value 'status: ERR'
}

