% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\docType{data}
\name{feeds}
\alias{feeds}
\title{Some RSS feeds.}
\format{A data frame with 10 rows and 2 variables:
\describe{
  \item{name}{nsame of the feed}
  \item{url}{url of the feed}
  }}
\source{
Author's personal choice.
}
\usage{
feeds
}
\description{
A dataset containing some RSS feeds chosen from the package author.
}
\keyword{datasets}

