% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distract.R
\name{links}
\alias{links}
\title{Return last entries retrieved.}
\usage{
links(entry.n = NULL)
}
\arguments{
\item{entry.n}{Index of the entry you want to retrieve the link.}
}
\value{
Links for each entry got from \code{distract()}
}
\description{
\code{distract(n)} returns a list of entries from the n-th feed in the feed
list. \code{links} provide the associated links (of the entire list or, if k
is given, of the k-th entry). When the links are then shown, you can use
\code{copy(k)} to copy the k-th entry in your system's clipboard, to be used
in your favourite browser.
}
\examples{
\dontrun{
links(1)
}
\dontrun{
links()
}
}

