% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distract.R
\name{remove.rss.feed}
\alias{remove.rss.feed}
\title{Remove a feed.}
\usage{
remove.rss.feed(feed.index)
}
\arguments{
\item{feed.index}{index of the feed to be removed}
}
\value{
"Removed" if success
}
\description{
Remove the n-th entry from the feed list.
}
\examples{
\dontrun{
remove.rss.feed(1)
}
}

