% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distract.R
\name{save.rss.feed}
\alias{save.rss.feed}
\title{Save a favourite feed in local env.}
\usage{
save.rss.feed(name, url)
}
\arguments{
\item{name}{Name of the feed}

\item{url}{URL}
}
\value{
"Saved" if succesfull, otherwise error message
}
\description{
If you want to add a url to the predefined list of feeds, pass it to this
function and it will be available to \code{distract()}.
}

