% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distract.R
\name{top.entries}
\alias{top.entries}
\title{Return most recent entries from a feed.}
\usage{
top.entries(url, number.of.entries = 5)
}
\arguments{
\item{url}{URL for the feed.}

\item{number.of.entries}{number of entries to return}
}
\value{
a dataframe with title of each entry (what) and its date / time
}
\description{
Given a url, it parses the feed and creates a dataframe, then it sorts by
date / time and return the most recent entries. It saves a complete dataframe
in local storage, and returns just title and date / time.
}
\examples{
\dontrun{top.entries("https://feeds.feedburner.com/RBloggers", 3)}
}

